Putting BicycleGenerator as seperate class
